Cleans up Binary and Unary Macros

+++ test/test.cpp
@@ -34,7 +34,7 @@ int main() {
     StealthTileMap::display(test && test2, "Test && Test2");
     StealthTileMap::display(test || test2, "Test || Test2");
     StealthTileMap::display(!test, "!Test");
-    StealthTileMap::display(StealthTileMap::transform<int, float, halfVal>(test), "Halved Int Test");
+    StealthTileMap::display(StealthTileMap::transform<float, int, halfVal>(test), "Halved Int Test");
 
     // Creating TileMap from array
     std::array<float, 4> init{{0, 1, 2, 3}};
